using express router on server instead of homebrew structure
