+++ routes/index.js
@@ -10,3 +10,5 @@ exports.templates = function (req, res) {
   var name = req.params.name;
   res.render('templates/' + name);
 };
+
+exports.postRouter =  require('./post.js');
